requirements.txt
# API 客户端
openai==1.3.5
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Web 框架
fastapi==0.104.1
uvicorn==0.24.0

# 任务队列
celery==5.3.6
redis==5.0.1
flower==2.0.1

# 工具库
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
pydantic==2.5.2
pydantic-settings==2.0.3

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 其他依赖
pydub==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2